requirements.txt
google-adk
python-telegram-bot
google-cloud-firestore
python-dotenv
httpx
